% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_node_perturbation.R
\name{find_node_perturbation}
\alias{find_node_perturbation}
\title{find_node_perturbation}
\usage{
find_node_perturbation(input_graph, how = 2, cycle = 1, limit = 0)
}
\arguments{
\item{input_graph}{The graph object that was processed with priming_graph function.}

\item{how}{The change of count (expression) of the given node in terms of fold change.}

\item{cycle}{The iteration of simulation.}

\item{limit}{The minimum fold change which can be taken into account for perturbation calculation on all nodes in terms of percentage.}
}
\value{
It gives a tibble form dataset that includes node names, perturbation efficiency and perturbed count of nodes.
}
\description{
calculates the mean of expression changes of all  nodes and finds the perturbed node count for all nodes in system.
}
\details{
find_node_perturbation calculates mean expression changes of elements after the change in the network in terms of percentage. It also calculates the number of nodes that have expression changes after the change occur in the network.
  The outputs of the function are the perturbation efficiency and perturbed count of nodes for each nodes.
}
\examples{

minsamp\%>\%
   priming_graph(competing_count = Competing_expression, miRNA_count = miRNA_expression, aff_factor = c(energy,seed_type), deg_factor = region)\%>\%
   find_node_perturbation(how = 3, cycle = 4)

}
