% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_how.R
\name{gene_knockdown}
\alias{gene_knockdown}
\title{Knocks down given node.}
\usage{
gene_knockdown(input_graph, node_name)
}
\arguments{
\item{input_graph}{The graph object that processed in previous step/s.}

\item{node_name}{The name of the node whose count is to be knocked down.}
}
\value{
the graph object.
}
\description{
Knocks down given node.
}
\details{
knocks down a given gene target.
}
\keyword{internal}
