% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{huge_example}
\alias{huge_example}
\title{huge example}
\format{A data frame with 7 variables and 26176 observation:
\describe{
  \item{competing}{name of gene}
  \item{miRNA}{name of miRNA}
  \item{competing_counts}{Expression values of competing element (gene)}
  \item{mirnaexpression_normal}{Expression value of miRNA elements in normal tissue}
  \item{Energy}{Energy of miRNA:target binding}
  \item{region_effect}{Coefficient for efficiency of location on target}
  \item{seed_type_effect}{Coefficient for efficiency of seed sequence of miRNA:target interaction}
  }}
\source{
Dataset was integrated by us.
}
\description{
A sample dataset which is utilised through integration of TCGA_E9_A1N5_normal,
    TCGA_E9_A1N5_mirnanormal and high-throughput experimental miRNA:gene dataset.
}
