% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_iteration.R
\name{iteration_graph}
\alias{iteration_graph}
\title{iteration_graph}
\usage{
iteration_graph(df, competing_count, miRNA_count, aff_factor = dummy,
  deg_factor = dummy, node_name, how, .iter = 100, limit = 0.2)
}
\arguments{
\item{df}{A data frame that includes the miRNA and competing targets.}

\item{competing_count}{The counts (expression) of competing elements of the dataset.}

\item{miRNA_count}{The counts (expression) of repressive element (miRNA) of the dataset.}

\item{aff_factor}{The parameter/s of binding between miRNA and targets.}

\item{deg_factor}{The parameter/s for degradation of bound miRNA:target complex.}

\item{node_name}{The name f node that is used as trigger.}

\item{how}{The change of count (expression) of the given node in terms of fold change.}

\item{.iter}{The maximum iteration prescribed by the user.}

\item{limit}{The minimum amount of change of any node.}
}
\value{
It gives the plot.
}
\description{
The function calculates the number of disturbed nodes for each iteration and than gives the plot that is ratio of iteration and affected node  in terms of percentage. .
}
\details{

}
\examples{

find_iteration(minsamp, Competing_expression, miRNA_expression, aff_factor = energy, deg_factor = region, node_name = "Gene2", how= 3)

}
