% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{midsamp}
\alias{midsamp}
\title{midsamp}
\format{A data frame with 7 variables and 26 observation of them:
\describe{
  \item{Genes}{symbol of gene}
  \item{miRNAs}{symol of miRNA}
  \item{Gene_Expression}{Expression values of competing gene}
  \item{miRNA_expression}{Expression value of miRNA}
  \item{seeds}{Coefficient for efficiency of seed type of miRNA:target interaction}
  \item{targeting_region}{Coefficient for efficiency of location on target}
  \item{Energy}{Energy of miRNA:target binding}
  }}
\source{
Dataset was created by us.
}
\description{
middle sized sample dataset
}
