% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{midsamp_new_counts}
\alias{midsamp_new_counts}
\title{midsamp_new_counts}
\format{A data frame with 4 variables and 26 observation of them:
\describe{
  \item{competing}{symbol of gene}
  \item{miRNA}{symol of miRNA}
  \item{Competing_count}{Expression values of competing gene}
  \item{miRNA_count}{Expression value of miRNA}
  }}
\source{
Dataset was created by us.
}
\description{
includes new expression values for middle sized sample dataset
}
