% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{minsamp}
\alias{minsamp}
\title{minsamp}
\format{A data frame with 7 variables and 7 observation of them:
\describe{
  \item{competing}{symbol of gene}
  \item{miRNA}{symol of miRNA}
  \item{Competing_expression}{Expression values of competing gene}
  \item{miRNA_expression}{Expression value of miRNA}
  \item{seed_type}{Coefficient for efficiency of seed sequence of miRNA:target interaction}
  \item{region}{Coefficient for efficiency of location on target}
  \item{energy}{Energy of miRNA:target binding}
  }}
\source{
Dataset was created by us.
}
\description{
minimal sample dataset
}
