% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mirtarbasegene}
\alias{mirtarbasegene}
\title{mirtarbasegene}
\format{Classes tbl_df, tbl and data.frame with 380627 observation of 2 variables:
\describe{
  \item{miRNA}{miRNA symbol}
  \item{Target}{target gene symbol}
  }}
\source{
\url{http://mirtarbase.mbc.nctu.edu.tw/php/index.php}
}
\description{
the dataset that includes miRNA:target gene interactions downloaded from mirtarbase
}
