% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{new_counts}
\alias{new_counts}
\title{new_counts}
\format{A data frame with 7 variables and 7 observation of them:
\describe{
  \item{Competing}{symbol of gene}
  \item{miRNA}{symol of miRNA}
  \item{Competing_count}{Expression values of competing gene}
  \item{miRNA_count}{Expression value of miRNA}
  }}
\source{
Dataset was created by us.
}
\description{
includes new expression values for minimal sample dataset
}
