% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_nodes.R
\name{prepare_rhs}
\alias{prepare_rhs}
\title{prepare_rhs}
\usage{
prepare_rhs(input_graph)
}
\arguments{
\item{input_graph}{Processed graph object in previous step.}
}
\description{
Carries the variables from edge to node.
}
\details{
The function is a helper function for processing of graph object with update_nodes function.
}
