% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(input_graph, cycle = 1)
}
\arguments{
\item{input_graph}{The graph object that processed in previous steps.}

\item{cycle}{Optimal iteration number for gaining steady-state.}
}
\value{
The graph.
}
\description{
simulate function uses the change in expression value/s as triggering.
}
\details{
The steady-state conditions of the system are disturbed after the change in the graph (with update_how or update_variables). In this case, the system tend to be steady state again. The arrangement of competetive profiles of the targets continue until all nodes are updated and steady-state nearly.
}
\examples{

priming_graph(minsamp, Competing_expression, miRNA_expression, aff_factor = c(seed_type,energy), deg_factor = region)\%>\%
   update_nodes(once = TRUE)\%>\%
   update_variables(new_counts)\%>\%
   simulate(cycle = 3)

}
