% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_knockdown}
\alias{simulate_knockdown}
\title{Simulates the system a gene is knocked down.}
\usage{
simulate_knockdown(input_graph, cycle = 1, node_name, knockdown = TRUE)
}
\arguments{
\item{input_graph}{The graph object that processed in previous steps.}

\item{cycle}{Optimal iteration number for gaining steady-state.}

\item{node_name}{The name of the node whose count is to be knocked down.}

\item{knockdown}{default is TRUE}
}
\value{
The graph.
}
\description{
Simulates the system a gene is knocked down.
}
\details{
The steady-state conditions of the system are disturbed with gene_knockdown() function. As a result of this, efficiencies of microRNAs in the network must be regulated.
}
\examples{

data("minsamp")

minsamp \%>\%
 priming_graph(competing_count = Competing_expression,
               miRNA_count = miRNA_expression)\%>\%
 gene_knockdown("Gene4", knockdown = TRUE)\%>\%
 simulate_knockdown(cycle = 2, "Gene4", knockdown = TRUE)

}
