% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TCGA_E9_A1N5_mirnanormal}
\alias{TCGA_E9_A1N5_mirnanormal}
\title{TCGA_E9_A1N5_mirnanormal}
\format{Classes tbl_df, tbl and data.frame with	750 observation of  6 variables:
\describe{
  \item{barcode}{Sample, normal tissue, barcode of patient based on TCGA}
  \item{mirbase_ID}{mirbase id of miRNA}
  \item{miRNA}{miRNA name}
  \item{Precusor}{Precusor id of miRNA which is given in miRNA variable}
  \item{total_read}{total reading count of miRNA which is produced from different gene locations}
  \item{total_RPM}{total RPM (reading per million) of miRNA}
  }}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\description{
The dataset contains mirna expression values for normal tissue sample of TCGA-E9-A1N5 barcoded patient
}
