% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TCGA_E9_A1N5_mirnatumor}
\alias{TCGA_E9_A1N5_mirnatumor}
\title{TCGA_E9_A1N5_mirnatumor}
\source{
TCGA (The Cancer Genome Atlas)
}
\description{
contains mirna expression values for tumor tissue sample of TCGA_E9_A1N5 barcoded patient
}
