% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TCGA_E9_A1N5_tumor}
\alias{TCGA_E9_A1N5_tumor}
\title{TCGA_E9_A1N5_tumor}
\source{
TCGA (The Cancer Genome Atlas)
}
\description{
contains gene expression values for cancer tissue sample of TCGA_E9_A1N5 barcoded patient
}
