% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_nodes.R
\name{update_nodes}
\alias{update_nodes}
\title{Carries variables from edge to node.}
\usage{
update_nodes(input_graph, once = FALSE, limit = 0)
}
\arguments{
\item{input_graph}{Processed graph object in previous step.}

\item{once}{The argument is about when the carrying process runs (internal use only)}

\item{limit}{absolute minimum amount of change required to be considered as up/down regulated element}
}
\value{
the graph object.
}
\description{
This function carries variables from edge to node and should be used
after `update_how` or `update_variables` functions
}
\details{
If the carrying process performs after priming_graph function, the argument must be TRUE.
    The function helps to visualisation of processed graph object, especially that includes too many nodes.This step makes it easily to follow the processes.
}
\examples{

data("minsamp")

minsamp \%>\%
  priming_graph(Competing_expression, miRNA_expression) \%>\%
  update_how("Gene2",2)

}
