% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priming_graph.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{The variable name that is normalized.}
}
\description{
normalizes the values according to maximum values inside a group. The helper function of priming_graph.
}
