% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TCGA_E9_A1N5_mirnanormal}
\alias{TCGA_E9_A1N5_mirnanormal}
\title{TCGA_E9_A1N5_mirnanormal}
\source{
TCGA (The Cancer Genome Atlas)
}
\description{
contains mirna expression values for normal tissue sample of TCGA_E9_A1N5 barcoded patient
}
