% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_nodes.R
\name{update_nodes}
\alias{update_nodes}
\title{update_nodes}
\usage{
update_nodes(input_graph, once = FALSE)
}
\arguments{
\item{input_graph}{Processed graph object in previous step.}

\item{once}{The argument is about when the carrying process runs.}
}
\value{
the graph object.
}
\description{
The function carries variables from edge to node.
}
\details{
If the carrying process performs after priming_graph function, the argument must be TRUE.
    The function helps to visualisation of processed graph object, especially that includes too many nodes.This step makes it easily to follow the processes.
}
\examples{

minsamp\%>\%
priming_graph(Competing_expression, miRNA_expression, aff_factor = c(seed_type,energy), deg_factor = region)\%>\%
   update_nodes(once = TRUE)

}
